Refactor silo component files
