+++ src/components/silo/views/Create.tsx
@@ -2,35 +2,23 @@
 
 import Button from "@/components/ui/Button"
 import Input from "@/components/ui/Input"
-import { archive, getByUid, update } from "@/lib/silo"
-import { updateSchema } from "@/lib/validation/silo"
-import { Silo } from "@/types/silo"
+import { useAuth } from "@/hooks/useAuth"
+import { create } from "@/lib/silo"
+import { createSchema } from "@/lib/validation/silo"
 import Link from "next/link"
 import { useRouter } from "next/navigation"
-import { ChangeEvent, FormEvent, useEffect, useState } from "react"
+import { ChangeEvent, FormEvent, useState } from "react"
 
-export default function SiloEditForm({ uid }: { uid: string }) {
+export default function SiloViewCreate() {
   const [formData, setFormData] = useState({
     name: "",
     description: "",
   })
-  const [silo, setSilo] = useState<Silo | null>(null)
   const [error, setError] = useState<string | null>(null)
   const [success, setSuccess] = useState<string | null>(null)
   const [isSubmitting, setIsSubmitting] = useState(false)
+  const { currentUser } = useAuth()
   const router = useRouter()
-
-  useEffect(() => {
-    const fetchSilo = async () => {
-      const silo = await getByUid(uid)
-      setFormData({
-        name: silo?.name ?? "",
-        description: silo?.description ?? "",
-      })
-      setSilo(silo)
-    }
-    fetchSilo()
-  }, [uid])
 
   const handleInputChange = (e: ChangeEvent<HTMLInputElement>) => {
     const { name, value } = e.target
@@ -44,14 +32,14 @@ export default function SiloEditForm({ uid }: { uid: string }) {
     setIsSubmitting(true)
 
     if (validateSchema()) {
-      await updateSilo()
+      await createSilo()
     }
 
     setIsSubmitting(false)
   }
 
   const validateSchema = (): boolean => {
-    const result = updateSchema.safeParse(formData)
+    const result = createSchema.safeParse(formData)
     if (result.error) {
       setError(result.error.errors[0].message)
       return false
@@ -59,17 +47,20 @@ export default function SiloEditForm({ uid }: { uid: string }) {
     return true
   }
 
-  const updateSilo = async () => {
+  const createSilo = async () => {
     try {
-      const result = await update(uid, {
-        name: formData.name,
-        description: formData.description,
-      })
+      const result = await create(
+        {
+          name: formData.name,
+          description: formData.description,
+        },
+        currentUser?.uid ?? "",
+      )
       if (result === null) {
         setError("Firebase foutmelding (zie console)")
         return
       }
-      setSuccess("Silo aangepast. Even geduld...")
+      setSuccess("Silo aangemaakt. Even geduld...")
       router.push("/silo")
     } catch (error: unknown) {
       if (error instanceof Error) {
@@ -80,15 +71,10 @@ export default function SiloEditForm({ uid }: { uid: string }) {
     }
   }
 
-  const archiveSilo = async () => {
-    await archive(silo?.uid ?? "")
-    router.push("/silo")
-  }
-
   return (
     <div className="w-96 space-y-2 text-center">
       <h1 className="text-center text-xl font-bold">
-        Pas de silo gegevens aan
+        Maak een nieuwe silo aan
       </h1>
 
       <form
@@ -109,24 +95,19 @@ export default function SiloEditForm({ uid }: { uid: string }) {
           value={formData.description}
           onChange={handleInputChange}
         />
-        <div className="flex items-center justify-between space-x-2 pt-2">
+
+        <div className="flex items-center justify-between space-x-2">
           <Link
             className="underline"
             href="/silo"
           >
             Annuleren
           </Link>
-          <div className="space-x-2">
-            <Button
-              label="Archiveren"
-              onClick={archiveSilo}
-            />
-            <Button
-              disabled={isSubmitting}
-              label="Silo aanpassen"
-              type="submit"
-            />
-          </div>
+          <Button
+            disabled={isSubmitting}
+            label="Silo aanmaken"
+            type="submit"
+          />
         </div>
 
         {error && <p className="text-red-500">{error}</p>}
